+++ CHANGELOG.md
@@ -10,3 +10,9 @@
 + Added result extension methods
 + Added type to notifications
 + Added notification type utility
+
+### V0.1.2
+
++ Added quick access customisable 'Success' boolean on results
++ Added missing clone elements for HttpResult
++ **BREAKING:** Moved trace out of notifications to the result